Project 4 Dictionaries
Created a dictionary as “a key-value pair collection,” with Boxer as an object. I utilized the document.getElementById("Dictionary").innerHTML method to output a value. Aslo, I used the deleted operator to remove a value from my dictionary and ensured the deleted value was specified by the document.getElementById("Dictionary").innerHTML method.