Project 6 Ternary Operators & Constructors
Wrote a function using HTML and JavaScript that utilizes a ternary operation using input from the browser.

Wrote a constructor function that uses the “new” and “this” keywords.

Created a function to display the results of the constructor function within an HTML element.

Wrote a nested function.

Write comments throughout your JavaScript explaining your code